# server/swagger.json
{"swagger":"2.0","info":{"description":"API created for Infinity app by CS Capstone team, Portland State University Fall/Winter 2017. \n\nThis document provides explanation and demonstration of available APIs for Infinity app. Those APIs conform to the RESTful design principles, so you can interact with the resources exposed via the API by passing parameters and HTTP verbs (GET, POST, PUT, and DELETE). For each API, it has general description, parameter requirements and a sample output.  Besides, the organization of this document also conforms to the design of the database.\n","version":"3.0.0","title":"uLab Infinity API","contact":{"name":"Universal Design Lab","url":"http://www.universaldesignlab.com/","email":"sennott@pdx.edu"},"license":{"name":"MIT License","url":"https://github.com/UniversalDesignLab/uLabCapstone/blob/master/LICENSE.md"}},"tags":[{"name":"Grids","description":"Data by grids"},{"name":"Lists","description":"Data by lists"},{"name":"Words","description":"Data by words"}],"schemes":["https"],"paths":{"/words":{"get":{"tags":["Words"],"summary":"Returns all words","operationId":"getAllWords","description":"Returns all available words","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/words"}}}}},"post":{"tags":["Words"],"summary":"Creates a new word for a list","operationId":"createWord","description":"Creates a new word for a specified list","consumes":["application/json"],"parameters":[{"name":"body","in":"body","required":true,"schema":{"type":"object","properties":{"name":{"type":"string"},"path":{"type":"string"},"text":{"type":"string"},"list":{"type":"string"},"grid":{"type":"string"}},"required":["name","path","text","list","grid"],"example":{"name":"love","path":"img/love.png","text":"love symbol","list":"verb","grid":"core vocabulary"}}}],"responses":{"201":{"description":"Created"}}}},"/grids":{"get":{"tags":["Grids"],"summary":"Returns all grids","description":"Returns grid_id and grid_title for all grids in database","responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/grids"}}}}}},"/grids/title/{grid_title}/words":{"get":{"tags":["Grids"],"summary":"Returns all words in a grid","operationId":"getAllWordsByGridName","description":"Returns all list and word data for a grid by grid title","parameters":[{"name":"grid_title","in":"path","description":"The title of the grid","type":"string","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"All grid data sent","schema":{"type":"array","items":{"$ref":"#/definitions/gridwords"}}}}}},"/grids/id/{grid_id}/words":{"get":{"tags":["Grids"],"summary":"Returns all words in a grid","operationId":"getAllWordsByGridId","description":"Returns all lists for a grid by grid_id","parameters":[{"name":"grid_id","in":"path","description":"The id of the grid","type":"integer","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/gridwords"}}}}}},"/grids/id/{grid_id}":{"get":{"tags":["Grids"],"summary":"Returns all lists in a grid","operationId":"getAllListsByGridId","description":"Returns all lists for a grid by grid_id","parameters":[{"name":"grid_id","in":"path","description":"The id of the grid","type":"integer","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/gridlists"}}}}}},"/grids/{grid_id}/lists/{list_id}":{"get":{"tags":["Grids"],"summary":"Returns all words in a grid by list","operationId":"getAllListWordsByListId","description":"Returns all words by grid_id and list_id","parameters":[{"name":"grid_id","in":"path","description":"The id of the grid","type":"integer","required":true,"format":"int32"},{"name":"list_id","in":"path","description":"the id of the list","type":"integer","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/gridwords"}}}}}},"/lists":{"post":{"tags":["Lists"],"summary":"Creates a new list","operationId":"createList","description":"Adds a new list to the database","consumes":["application/json"],"parameters":[{"name":"body","in":"body","required":true,"schema":{"type":"object","properties":{"title":{"type":"string"}},"required":["title"],"example":{"title":"my list"}}}],"responses":{"201":{"description":"Created","schema":{"example":{"status":"success","message":"New list my list created","data":{"id":9}}}}}}},"/lists/title/{list_title}":{"get":{"tags":["Lists"],"summary":"Returns all words in a list","operationId":"getAllListsByName","description":"Returns words of a list by list_title","parameters":[{"name":"list_title","in":"path","description":"The title of the list","type":"string","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/lists"}}}}}},"/lists/id/{list_id}":{"get":{"tags":["Lists"],"summary":"Returns all words in a list","operationId":"getAllListsById","description":"Returns words of a list by list_id","parameters":[{"name":"list_id","in":"path","description":"The id of the list","type":"integer","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/lists"}}}}}},"/words/id/{word_id}":{"get":{"tags":["Words"],"summary":"Returns word data","operationId":"getWordById","description":"Returns word data by word_id","parameters":[{"name":"word_id","in":"path","description":"The id of the word","type":"integer","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","items":{"$ref":"#/definitions/words"}}}}}},"/words/name/{word_name}":{"get":{"tags":["Words"],"summary":"Returns word data","operationId":"getWordByName","description":"Returns word data by word_name","parameters":[{"name":"word_name","in":"path","description":"The name of the word","type":"string","required":true}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","items":{"$ref":"#/definitions/words"}}}}}},"/words/list_id/{list_id}/word_id/{word_id}":{"delete":{"tags":["Words"],"summary":"Deletes a word from a list","operationId":"deleteWordByID","description":"Deletes a word from a list but remains in the database for future reuse","parameters":[{"name":"list_id","in":"path","description":"The id of the list the word is to be removed from","type":"integer","required":true,"format":"int32"},{"name":"word_id","in":"path","description":"The id of the word to be deleted","type":"integer","required":true,"format":"int32"}],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"example":{"status":"success","message":"Removed 1 word"}}}}}},"/imgupload":{"post":{"tags":["File Upload"],"summary":"Uploads an image file.","operationId":"createImage","description":"Adds a new image to client/public/img.  Allows only .png files with max size a 10MB","consumes":["multipart/form-data"],"parameters":[{"name":"userfile","in":"formData","description":"file to upload","required":true,"type":"file"}],"produces":["application/json"],"responses":{"201":{"description":"Created","schema":{"example":{"status":"success","message":"File uploaded"}}}}}}},"definitions":{"words":{"type":"object","required":["word_id","word","symbol_path","symbol_text"],"properties":{"word_id":{"type":"integer","format":"int32","example":1},"word":{"type":"string","example":"think"},"symbol_path":{"type":"string","example":"/img/think.png"},"symbol_text":{"type":"string","example":"think symbol"}}},"grids":{"type":"object","required":["grid_id","grid_title"],"properties":{"grid_id":{"type":"integer","format":"int32","example":1},"grid_title":{"type":"string","example":"core vocabulary"}}},"gridlists":{"type":"object","required":["grid_id","list_id","list_title"],"properties":{"grid_id":{"type":"integer","format":"int32","example":1},"list_id":{"type":"integer","format":"int32","example":1},"list_title":{"type":"string","example":"verb"}}},"lists":{"type":"object","required":["list_id","list_title","word_id","word","symbol_id","symbol_path","symbol_text"],"properties":{"list_id":{"type":"integer","format":"int32","example":1},"list_title":{"type":"string","example":"verb"},"word_id":{"type":"integer","format":"int32","example":1},"word":{"type":"string","example":"think"},"symbol_id":{"type":"integer","format":"int32","example":1},"symbol_path":{"type":"string","example":"/img/think.png"},"symbol_text":{"type":"string","example":"think symbol"}}},"gridwords":{"type":"object","required":["grid_id","grid_title","list_id","list_title","word_id","word","symbol_id","symbol_name","symbol_path","symbol_text"],"properties":{"grid_id":{"type":"integer","format":"int32","example":1},"grid_title":{"type":"string","example":"core vocabulary"},"list_id":{"type":"integer","format":"int32","example":1},"list_title":{"type":"string","example":"verb"},"word_id":{"type":"integer","format":"int32","example":1},"word":{"type":"string","example":"think"},"symbol_id":{"type":"integer","format":"int32","example":1},"symbol_name":{"type":"string","example":"think"},"symbol_path":{"type":"string","example":"/img/think.png"},"symbol_text":{"type":"string","example":"think symbol"}}}},"host":"virtserver.swaggerhub.com","basePath":"/uLabCapstone/Infinity/1.0.0"}